Finished rewrite, improved local card choice.. fixed some async functions. now improve stability
